Python - Simulation, measurement/requirements.txt
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
PyAutoGUI==0.9.54
pygame==2.6.1
pywin32==308
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.14.1
screeninfo==0.8.1
